+++ src/home/manage-project/ProjectCard.tsx
@@ -1,4 +1,4 @@
-import { Button } from "@/components/ui/button";
+import { Button } from "@/component/ui/button";
 import { ImageIcon, Pencil, Trash2 } from "lucide-react";
 
 interface ProjectCardProps {
@@ -26,12 +26,12 @@ export default function ProjectCard({
 }: ProjectCardProps) {
 
   return (
-    <div className="flex border border-border bg-card rounded-lg p-6 gap-4 justify-between">
-      <div className="flex gap-4">
+    <div className="flex flex-col lg:flex-row border border-border bg-card rounded-lg p-6 gap-4 justify-between w-full">
+      <div className="flex flex-col lg:flex-row gap-4 w-auto">
         <span>
           {image === '' ? 
             <div>
-              <div className="w-56 h-36 rounded-lg bg-muted border-2 border-border flex items-center justify-center">
+              <div className="w-auto lg:w-56 h-36 lg:h-36 rounded-lg bg-muted border-2 border-border flex items-center justify-center">
                 <ImageIcon className="w-8 h-8" />
               </div>
             </div> 
@@ -47,7 +47,7 @@ export default function ProjectCard({
             <p className="text-sm text-muted-foreground">
               Link:{" "}
               <a href={link} target="_blank" rel="noopener noreferrer" className="text-primary underline">
-                {link}
+                {link.slice(0, 30)}...
               </a>
             </p>
             <p className="text-sm text-muted-foreground">
@@ -59,12 +59,12 @@ export default function ProjectCard({
         </span>
       </div>
 
-      <div className="flex gap-1">
-        <Button variant="ghost" onClick={() => handleEdit(id)}>
+      <div className="grid-cols-2 grid lg:flex gap-2">
+        <Button variant="outline" onClick={() => handleEdit(id)} className="col-span-1">
           <Pencil className="w-4 h-4" />
           Edit
         </Button>
-        <Button variant="ghost" onClick={() => handleDelete(id)}>
+        <Button variant="destructive" onClick={() => handleDelete(id)} className="col-span-1">
           <Trash2 className="w-4 h-4" />
           Delete
         </Button>