+++ CHANGELOG.md
@@ -5,11 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 0.1.1 - 11/04/2025
+## [0.1.2] - 11/05/2025
+
+### Added
+
+- Functional Manage Employment tab
+- Added animation on card pop ups, hover, to remove static feel
+- Added DELETE function on skills, projects, and employments
+- Mobile-responsiveness and layout changes
+
+## [0.1.1] - 11/04/2025
 
 ### Added
 
 - Added Middleware and Authentication
-- Manage Profile and Manage Projects
+- Functional Manage Profile and Manage Projects tab
 - Collapsible Sidebar
 - Dark Mode/Light Mode